+++ CHANGES.md
@@ -3,6 +3,7 @@ Change Log
 - Version 2.0.2 (Pending)
     - Split out badge type color bar and age range color bar
     - Added badge type to Print Test Badge screen
+    - Added .json file import for staff information
 
 - Version 2.0.1 (7/4/2017)
     - Fixed bug where adding a minor without a parental consent form would display red "!" instead